backend/requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
python-multipart==0.0.9
pydantic==2.6.4
openai==1.14.2
python-dotenv==1.0.1
